pyproject.toml
[project]
name = "razorpay-jshop"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-bootstrap5>=26.1",
    "django-extensions>=4.1",
    "django-stubs>=5.2.9",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "razorpay>=2.0.0",
]
